[MOVE/#11] MyTodo, OurTodo 패키지 분리
