+++ presentation/src/main/java/com/going/presentation/todo/ourtodo/OurTodoFragment.kt
@@ -1,4 +1,4 @@
-package com.going.presentation.todo
+package com.going.presentation.todo.ourtodo
 
 import android.os.Bundle
 import android.text.SpannableStringBuilder
